+++ commands/info.js
@@ -5,6 +5,8 @@ const { prefix, default_color } = require('../config.json');
 module.exports = {
     name: 'info',
     description: 'Provides information about the bot itself.',
+    cooldown: 1,
+    guildOnly: false,
     execute(message) {
       message.channel.send("Here you go.", {
         embed: {