+++ commands/quote.js
@@ -5,6 +5,8 @@ const { default_color } = require("../config.json");
 module.exports = {
     name: "quote",
     description: "Sends a quote from Jotaro.",
+    cooldown: 1,
+    guildOnly: false,
     execute(message) {
         var getArc = functions.getRandom(quotes.series[0].arc.length);
         var obj = quotes.series[0].arc[getArc];
@@ -15,7 +17,6 @@ module.exports = {
         
         var randTwo = functions.getRandom(obj[arcName].length);
         var quoteObj = obj[arcName][randTwo];
-        console.log(quoteObj.quote);
         
         message.channel.send("", {
             embed: {