+++ commands/reload.js
@@ -3,6 +3,8 @@ const { owner_ID } = require('../config.json');
 module.exports = {
     name: 'reload',
     description: 'Reloads commands.',
+    cooldown: 0,
+    guildOnly: false,
     args: true,
     execute(message, args) {
       if (message.author.id !== owner_ID) return;